requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

# Azure Functions framework
azure-functions==1.18.0

applicationinsights==0.11.10

# Azure Blob Storage SDK for backup container functionality
azure-storage-blob==12.19.0

# Utilities and helpers
backoff==2.2.1
requests==2.32.3
cryptography==43.0.1